Fix and simplify grader
